+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 1.0.2601.172
+## Added
+- Added "Power on" request that useful for devices powered from mains but has a backup-battery
+- Added idf_component to be able to use PowerManagement from IDF components registry
+
 # 1.0.2601.171
-# First release
+## First release